fixed a bunch of compiler warnings

+++ c++/2021/day-1.hpp
@@ -8,7 +8,7 @@ const auto input = extractNumbers<uint16_t>(aoc::getInput(2021, 1));
 
 uint64_t part1() {
     int count = 0;
-    for(int i = 1; i < input.size(); i++) {
+    for(size_t i = 1; i < input.size(); i++) {
         count += input[i] > input[i - 1];
     }
     return count;
@@ -16,7 +16,7 @@ uint64_t part1() {
 
 uint64_t part2() {
     int count = 0;
-    for(int i = 3; i < input.size(); i++) {
+    for(size_t i = 3; i < input.size(); i++) {
         int last = input[i - 3] + input[i - 2] + input[i - 1];
         int current = input[i - 2] + input[i - 1] + input[i];
         count += current > last;
